# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: Academia El Soto API
  description: REST API for Academia El Soto
  contact:
    name: Grupo 14
  version: 1.0.0
servers:
- url: https://localhost:8443
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/courses/{id}:
    get:
      tags:
      - course-rest-controller
      operationId: getCourseById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CourseResponseDto"
    put:
      tags:
      - course-rest-controller
      operationId: updateCourse
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseRequestDto"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CourseResponseDto"
    delete:
      tags:
      - course-rest-controller
      operationId: deleteCourse
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
...
  /api/v1/courses:
    get:
      tags:
      - course-rest-controller
      operationId: getCourses
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCourseResponseDto"
    post:
      tags:
      - course-rest-controller
      operationId: createCourse
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseRequestDto"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CourseResponseDto"
  /api/v1/courses/{id}/image:
    get:
      tags:
      - course-rest-controller
      operationId: downloadCourseImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: binary
    post:
      tags:
      - course-rest-controller
      operationId: uploadCourseImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/carts/{cartId}/courses/{courseId}:
    post:
      tags:
      - cart-rest-controller
      operationId: addCourseToCart
      parameters:
      - name: cartId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: courseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponseDto"
    delete:
      tags:
      - cart-rest-controller
      operationId: removeCourseFromCart
      parameters:
      - name: cartId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: courseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponseDto"
  /api/v1/courses/{id}/recommended-books:
    get:
      tags:
      - course-rest-controller
      operationId: getRecommendedBooks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
  /api/v1/courses/{id}/comments:
    get:
      tags:
      - course-rest-controller
      operationId: getCourseComments
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentResponseDto"
...
  /api/v1/courses/{courseId}/image:
    delete:
      tags:
      - course-rest-controller
      operationId: deleteCourseImage
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
...
components:
  schemas:
...
    CourseRequestDto:
      type: object
      properties:
        courseName:
          type: string
          maxLength: 100
          minLength: 0
        teacher:
          type: string
          maxLength: 100
          minLength: 0
        price:
          type: integer
          format: int32
          minimum: 0
        description:
          type: string
          maxLength: 2000
          minLength: 0
        students:
          type: integer
          format: int32
          minimum: 0
        imageIds:
          type: array
          items:
            type: integer
            format: int64
      required:
      - courseName
      - description
      - price
      - students
      - teacher
    CourseResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        courseName:
          type: string
        teacher:
          type: string
        price:
          type: integer
          format: int32
        description:
          type: string
        students:
          type: array
          items:
            $ref: "#/components/schemas/UserBasicDto"
        imageIds:
          type: array
          items:
            type: integer
            format: int64
    UserBasicDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userName:
          type: string
        email:
          type: string
...
    CommentResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        user:
          type: string
        publicationDate:
          type: string
        courseId:
          type: integer
          format: int64
        courseName:
          type: string
...
    CartResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        product:
          type: string
        price:
          type: integer
          format: int32
        userName:
          type: string
        courses:
          type: array
          items:
            $ref: "#/components/schemas/CourseResponseDto"
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageCourseResponseDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CourseResponseDto"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT